requirements.txt
groq>=0.9.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
pymupdf>=1.24.0
gradio>=4.40.0
python-dotenv>=1.0.0